+++ ios-device-lib/Classes/Terminals/BBPOS/Model/BBPOSDevice.swift
@@ -1,0 +1,11 @@
+//
+//  BBPOSDevice.swift
+//  ios-device-lib
+//
+
+import Foundation
+
+struct BBPOSDevice: Codable {
+    let identifier: String
+    let name: String
+}