+++ ios-device-lib/Classes/Terminals/Utility/GlobalMobileSDK/Protocols/GMSDeviceReconnectionInterface.swift
@@ -1,0 +1,11 @@
+//
+//  GMSDeviceReconnectionInterface.swift
+//  ios-device-lib
+//
+
+import Foundation
+
+protocol GMSDeviceReconnectionInterface {
+    func hasSavedDevice(_ terminalType: TerminalType) -> Bool
+    func reconnectLastDevice(_ terminalType: TerminalType, connectingFinishBlock : @escaping (Bool?) -> Void)
+}